test: isolate Prometheus runtime identity coverage
